+++ frontend/src/app/page.tsx
@@ -1,4 +1,4 @@
-import Signup from "./components/pages/Signup";
+import Signup from "../components/pages/Signup";
 
 export default function Home() {
   return <Signup />;
